Move CategoryMerchandising util to a CategoryMerchandisingResult
