+++ src/Result/Graphql/Recommendation/CategoryMerchandisingResult.php
@@ -38,6 +38,13 @@ namespace Nosto\Result\Graphql\Recommendation;
 
 class CategoryMerchandisingResult
 {
+    const DISPATCH_EVENT_NAME_POST_RESULTS = 'nosto_post_cmp_results';
+    const DISPATCH_EVENT_NAME_PRE_RESULTS = 'nosto_pre_cmp_results';
+    const DISPATCH_EVENT_KEY_REQUEST = 'categoryMerchandising';
+    const DISPATCH_EVENT_KEY_RESULT = 'result';
+    const DISPATCH_EVENT_KEY_LIMIT = 'limit';
+    const DISPATCH_EVENT_KEY_PAGE = 'page';
+
     /** @var ResultSet $resultSet */
     private $resultSet;
 
@@ -99,5 +106,20 @@ class CategoryMerchandisingResult
     public function getBatchToken()
     {
         return $this->batchToken;
+    }
+
+    /**
+     * @param CategoryMerchandisingResult $result
+     * @return array
+     */
+    public static function parseProductIds(CategoryMerchandisingResult $result)
+    {
+        $productIds = [];
+        foreach ($result->getResultSet() as $item) {
+            if ($item->getProductId() && is_numeric($item->getProductId())) {
+                $productIds[] = $item->getProductId();
+            }
+        }
+        return $productIds;
     }
 }
